+++ serde_arrow/src/internal/mod.rs
@@ -1,3 +1,4 @@
+pub mod array_builder;
 pub mod common;
 pub mod decimal;
 pub mod deserialization;
@@ -5,9 +6,10 @@ pub mod deserializer;
 pub mod dsl;
 pub mod error;
 pub mod event;
-pub mod generic;
 pub mod schema;
 pub mod serialization;
+pub mod serializer;
 pub mod sink;
 pub mod source;
 pub mod tracing;
+pub mod utils;