+++ serde_arrow/src/test_end_to_end/mod.rs
@@ -4,3 +4,4 @@ mod issue_137_schema_like_from_arrow_schema;
 mod issue_90;
 mod test_docs_examples;
 mod test_items;
+mod test_serialize_deserialize;