+++ serde_arrow/src/test_impls/examples.rs
@@ -1,5 +1,3 @@
-use super::macros::*;
-
 use super::utils::Test;
 use crate::{
     internal::schema::{GenericDataType, GenericField},
@@ -383,45 +381,56 @@ fn issue_57() {
         .check_nulls(&[&[false], &[false], &[false], &[false]]);
 }
 
-test_roundtrip_arrays!(
-    simple_example {
-        #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
-        struct S {
-            a: f32,
-            b: u32,
-        }
+#[test]
+fn simple_example() {
+    #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
+    struct S {
+        a: f32,
+        b: u32,
+    }
 
-        let items = &[
-            S{ a: 2.0, b: 4 },
-            S{ a: -123.0, b: 9 },
-        ];
-        let fields = &[
+    let items = &[S { a: 2.0, b: 4 }, S { a: -123.0, b: 9 }];
+
+    Test::new()
+        .with_schema(vec![
             GenericField::new("a", GenericDataType::F32, false),
             GenericField::new("b", GenericDataType::U32, false),
-        ];
+        ])
+        .trace_schema_from_samples(items, TracingOptions::default().allow_null_fields(true))
+        .serialize(items)
+        .deserialize(items)
+        .check_nulls(&[&[false, false], &[false, false]]);
+}
+
+#[test]
+fn top_level_nullables() {
+    #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
+    struct S {
+        a: Option<f32>,
+        b: Option<u32>,
     }
-    assert_round_trip(fields, items);
-);
 
-test_roundtrip_arrays!(
-    toplevel_nullables {
-        #[derive(PartialEq, Clone, Debug, Serialize, Deserialize)]
-        struct S {
-            a: Option<f32>,
-            b: Option<u32>,
-        }
+    let items = &[
+        S {
+            a: Some(2.0),
+            b: None,
+        },
+        S {
+            a: None,
+            b: Some(9),
+        },
+    ];
 
-        let items = &[
-            S{ a: Some(2.0), b: None },
-            S{ a: None, b: Some(9) },
-        ];
-        let fields = &[
+    Test::new()
+        .with_schema(vec![
             GenericField::new("a", GenericDataType::F32, true),
             GenericField::new("b", GenericDataType::U32, true),
-        ];
-    }
-    assert_round_trip(fields, items);
-);
+        ])
+        .trace_schema_from_samples(items, TracingOptions::default().allow_null_fields(true))
+        .serialize(items)
+        .deserialize(items)
+        .check_nulls(&[&[false, true], &[true, false]]);
+}
 
 #[test]
 fn new_type_wrappers() {