+++ Changes.md
@@ -1,5 +1,17 @@
 # Change log
 
+## 0.11.5
+
+- Add `serde_arrow::Serializer`
+- Add support for new type wrappers, tuples and tuple structs to
+  `serde_arrow::Deserializer`
+- Add a generic `serde_arrow::ArrayBuilder` with support for both `arrow` and
+  `arrow2`
+- Implement `TryFrom<&[Field]>` (`arrow` and `arrow2`) and
+  `TryFrom<&[FieldRef]>` (`arrow` only) for `SerdeArrowSchema`
+- Implement `TryFrom<&SerdeArrowSchema>` for `Vec<Field>` and `Vec<FieldRef>`
+  for `arrow`
+
 ## 0.11.4
 
 - Add `serde_arrow::Deserializer`